+++ wsgi_boost/response.h
@@ -25,6 +25,7 @@ namespace wsgi_boost
 	private:
 		const std::string m_server_name = "WsgiBoost v." WSGI_BOOST_VERSION;
 		Connection& m_connection;
+		bool m_header_sent = false;
 
 	public:
 		std::string http_version = "HTTP/1.1";
@@ -35,10 +36,20 @@ namespace wsgi_boost
 
 		explicit Response(Connection& connection) : m_connection{ connection } {}
 
+		// Send HTTP header (status code + headers)
 		boost::system::error_code send_header(const std::string& status, headers_type& headers, bool async = false);
 
+		// Send data to the client
 		boost::system::error_code send_data(const std::string& data, bool async = false);
 
+		// Send a plain text HTTP message to a client
 		boost::system::error_code send_mesage(const std::string& status, const std::string& message = std::string());
+
+		// Send a html HTTP message to a client
+		boost::system::error_code send_html(const std::string& status, const std::string& title,
+			const std::string& header, const std::string& text);
+
+		// Check if HTTP header has been sent
+		bool header_sent() { return m_header_sent;  }
 	};
 }