+++ wsgi_boost/wsgi_boost.cpp
@@ -16,7 +16,7 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 	PyEval_InitThreads(); // Initialize GIL
 
 	py::register_exception_translator<StopIteration>(&stop_iteration_translator);
-	py::register_exception_translator<RuntimeError>(&runtime_error_translator);
+	py::register_exception_translator<std::runtime_error>(&runtime_error_translator);
 	
 	py::scope module;
 	module.attr("__doc__") = "This module provides WSGI/HTTP server class";
@@ -31,7 +31,7 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 
 	py::class_<HttpServer, boost::noncopyable>("WsgiBoostHttp",
 
-		"WsgiBoostHttp(ip_address='', port=8000, num_threads=1)\n\n"
+		"WsgiBoostHttp(ip_address='', port=8000, threads=1)\n\n"
 
 		"PEP-3333-compliant multi-threaded WSGI server\n\n"
 
@@ -42,12 +42,14 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 		":type ip_address: str\n"
 		":param port: server's port\n"
 		":type port: int\n"
-		":param num_threads: the number of threads for the server to run\n"
-		":type num_threads: int\n\n"
+		":param threads: the number of threads for the server to run\n"
+		":type threads: int\n\n"
 
-		"Usage::\n\n"
+		"Usage:\n\n"
 
-		"	import wsgi_boost\n\n"
+		".. code-block:: python\n\n"
+
+		"	from wsgi_boost import WsgiBoostHttp\n\n"
 
 		"	def hello_app(environ, start_response):\n"
 		"		content = 'Hello World!'\n"
@@ -56,16 +58,16 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 		"		start_response(status, response_headers)\n"
 		"		return[content]\n\n"
 
-		"	httpd = wsgi_boost.WsgiBoostHttp(num_threads=4)\n"
+		"	httpd = WsgiBoostHttp(threads=4)\n"
 		"	httpd.set_app(hello_app)\n"
 		"	httpd.start()\n"
 		,
 
-		py::init<std::string, unsigned short, unsigned int>((py::arg("ip_address") = "", py::arg("port") = 8000, py::arg("num_threads") = 1)))
+		py::init<std::string, unsigned short, unsigned int>((py::arg("ip_address") = "", py::arg("port") = 8000, py::arg("threads") = 1)))
 
 		.add_property("is_running", &HttpServer::is_running, "Get server running status")
 
-		.def_readwrite("use_gzip", &HttpServer::use_gzip, "Use gzip compression for static content, default: ``False``")
+		.def_readwrite("use_gzip", &HttpServer::use_gzip, "Use gzip compression for static content, default: ``True``")
 
 		.def_readwrite("host_hame", &HttpServer::host_name, "Get or set the host name, default: automatically determined")
 
@@ -73,7 +75,7 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 			"Get or set timeout for receiving HTTP request headers\n\n"
 
 			"This is the max. interval for reciving request headers before closing connection.\n"
-			"Defaul: 5s"
+			"Default: 5s"
 			)
 
 		.def_readwrite("content_timeout", &HttpServer::content_timeout,
@@ -81,11 +83,11 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 
 			"This is the max. interval for receiving POST/PUT content\n"
 			"or sending response before closing connection.\n"
-			"Defaul: 300s"
+			"Default: 300s"
 			)
 
 		.def_readwrite("url_scheme", &HttpServer::url_scheme,
-			"Get os set url scheme -- http or https (Default: ``'http'``)"
+			"Get or set url scheme -- http or https (default: ``'http'``)"
 			)
 
 		.def_readwrite("static_cache_control", &HttpServer::static_cache_control,
@@ -123,7 +125,7 @@ BOOST_PYTHON_MODULE(wsgi_boost)
 			"Set a WSGI application to be served\n\n"
 
 			":param app: a WSGI application to be served as an executable object\n"
-			":raises: RuntimeError on attempt to set a WSGI application while the server is running"
+			":raises RuntimeError: on attempt to set a WSGI application while the server is running"
 		)
 		;