+++ Changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+1.0.0
+-----
+
+- Renamed ``num_threads`` constructor parameter to ``threads`` (**breaking change!**).
+- Now ``Transfer-Encoding: chunked`` is used if a WSGI application
+  does not provide ``Content-Length`` header.
+- Error messages in HTML.
+- Code optimization.
+
 0.9.8
 -----